backend/requirements.txt
fastapi==0.128.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.12.5
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
pillow==11.1.0
sqlalchemy-admin==0.1.3
